Переименовал переменные, отформатировал код, переименовал классы и пакеты, добавил javadoc
