--- src/main/java/Belousov/ConverterExeption.java
@@ -1,9 +1,0 @@
-package Belousov;
-
-public class ConverterExeption extends Exception {
-
-    public ConverterExeption (String message){
-
-        super(message);
-    }
-}
