--- src/main/java/Belousov/Repository.java
@@ -1,9 +1,0 @@
-package Belousov;
-
-public interface Repository<V> {
-
-    public V getEntity(int index);
-    public V[] getAllEntity();
-    public void setEntity(V entyty);
-    public void setAllEntity(V[] entytys);
-}
