--- src/main/java/Belousov/ServiceExeption.java
@@ -1,9 +1,0 @@
-package Belousov;
-
-public class ServiceExeption extends RuntimeException {
-
-    public ServiceExeption(String message){
-
-        super(message);
-    }
-}
