+++ src/main/java/com/epam/belousov/exeptions/ConverterExeption.java
@@ -1,0 +1,13 @@
+package com.epam.belousov.exeptions;
+
+public class ConverterExeption extends Exception {
+
+    /**
+     * Construct a ConverterExeption with exeption message
+     *
+     * @param message that could be logged and analysed
+     */
+    public ConverterExeption(String message) {
+        super(message);
+    }
+}
