+++ src/main/java/com/epam/belousov/service/Converter.java
@@ -1,0 +1,60 @@
+package com.epam.belousov.service;
+
+import com.epam.belousov.exeptions.ConverterExeption;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class Converter {
+
+    /**
+     * This method convert DTO to entity
+     *
+     * @param humanDto - entry DTO
+     * @return human entity or exeption
+     * @throws ConverterExeption in case of either entry parameter is null
+     */
+    public Human convertToHuman(HumanDto humanDto) throws ConverterExeption {
+        if (humanDto == null) {
+            throw new ConverterExeption("Конвертация не инициализированного экземпляра класса HumanDto");
+        }
+        Human human;
+        human = new Human(humanDto.getId(),
+                humanDto.getName(),
+                humanDto.getBrithDate().getYear(),
+                humanDto.getBrithDate().getMonth(),
+                humanDto.getBrithDate().getDay(),
+                humanDto.getAdress().getStreet(),
+                humanDto.getAdress().getHouseNumber(),
+                humanDto.getAdress().getFloorNumber(),
+                humanDto.getAdress().getFlatNumber());
+
+        log.info("Конвертация из DTO в Entity прошла успешно");
+        return human;
+    }
+
+    /**
+     * This method convert entity to DTO
+     *
+     * @param human - entry entity
+     * @return human DTO or exeption
+     * @throws ConverterExeption in case of either entry parameter is null
+     */
+    public HumanDto convertToDto(Human human) throws ConverterExeption {
+        if (human == null) {
+            throw new ConverterExeption("Конвертация не инициализированного экземпляра класса Human");
+        }
+        HumanDto humanDto;
+        humanDto = new HumanDto(human.getId(),
+                human.getName(),
+                human.getBrithDate().getYear(),
+                human.getBrithDate().getMonth(),
+                human.getBrithDate().getDay(),
+                human.getAdress().getStreet(),
+                human.getAdress().getHouseNumber(),
+                human.getAdress().getFloorNumber(),
+                human.getAdress().getFlatNumber());
+
+        log.info("Конвертация из Entity в DTO прошла успешно");
+        return humanDto;
+    }
+}
